// src/types.ts
export interface Organization {
  id: string;
  name: string;
  domain: string;
  peers: Peer[];
  mspID?: string;
  country?: string;
  state?: string;
  locality?: string;
  type?: string;
}

export interface Peer {
  id: string;
  name: string;
  port: number;
  status: 'pending' | 'running' | 'stopped';
  couchDBPort?: number;
  chaincodePort?: number;
}

export interface Orderer {
  id: string;
  name: string;
  domain: string;
  port: number;
  status: 'pending' | 'running' | 'stopped';
  type: 'solo' | 'etcdraft';
  batchTimeout?: string;
  batchSize?: {
    maxMessageCount: number;
    absoluteMaxBytes: number;
    preferredMaxBytes: number;
  };
}

export interface NetworkConfig {
  organizations: Organization[];
  orderers: Orderer[];
  channelName?: string;
  consortium?: string;
  networkVersion?: string;
  template?: NetworkTemplate;
  stateDatabase?: 'goleveldb' | 'CouchDB';
}

export interface YAMLConfig {
  configtx: string;
  cryptoConfig: string;
  dockerCompose: string;
}

export interface NetworkTemplate {
  id: string;
  name: string;
  description: string;
  channelName: string;
  organizations: {
    name: string;
    type: string;
    peerCount: number;
  }[];
  ordererCount: number;
  stateDatabase: 'goleveldb' | 'CouchDB';
}